Update mounts_checker method
There are select devices that fail searching for partitions using CRB's mount-checking methods. Update it with SuperR method and adapt it to how the toolkit prop works. Also add up_param detection.

This will also add a way for the updater-script to determine if the search fails and notify user for it.